Updated Cards with spinners
