+++ package.json
@@ -16,6 +16,7 @@
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
     "react-router-dom": "^6.8.0",
+    "react-spinners": "^0.13.8",
     "use-sound": "^4.0.1"
   },
   "devDependencies": {
